pyproject.toml
[tool.poetry]
name = "motionlink"
version = "0.1.0"
description = ""
authors = ["EugeneChicevic <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"
discord = "^1.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
